requirements_local.txt
flask==2.2.3
flask-login==0.6.2
flask-sqlalchemy==3.0.3
flask-wtf==1.1.1
email-validator==2.0.0
google-generativeai==0.3.0
jinja2==3.1.2
markupsafe==2.1.2
numpy==1.24.2
openai==1.3.0
openpyxl==3.1.2
pandas==2.0.0
psycopg2-binary==2.9.6
PyPDF2==3.0.1
python-dotenv==1.0.0
pytz==2023.3
sqlalchemy==2.0.7
trafilatura==1.6.0
werkzeug==2.3.0
wtforms==3.0.1
xhtml2pdf==0.2.9